server/requirements.txt
alembic==1.8.1
aniso8601==9.0.1
black==22.6.0
click==8.1.3
colorama==0.4.5
Flask==2.2.2
Flask-Migrate==3.1.0
Flask-RESTful==0.3.9
Flask-SQLAlchemy==2.5.1
greenlet==1.1.3
itsdangerous==2.1.2
Jinja2==3.1.2
ma==1.0.2
ma.mongo==1.0.2
Mako==1.2.1
MarkupSafe==2.1.1
marshmallow==3.17.1
mypy-extensions==0.4.3
packaging==21.3
password-strength==0.0.3.post2
pathspec==0.9.0
platformdirs==2.5.2
psycopg2-binary==2.9.3
pyparsing==3.0.9
python-decouple==3.6
python-dotenv==0.20.0
pytz==2022.2.1
six==1.16.0
SQLAlchemy==1.4.40
tomli==2.0.1
Werkzeug==2.2.2
